% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concentrations.R
\name{get_concentrations}
\alias{get_concentrations}
\title{Get Table of Normal Concentrations}
\usage{
get_concentrations(x)
}
\arguments{
\item{x}{a numeric value of the HMDB accession ID (ie, 000001)}
}
\value{
a \code{tbl_df} of Normal Concentration information
}
\description{
Retrieve the \verb{Normal Concentrations} table for an accession entry from the Human Metabolome Database (HMDB)
}
